test: Use [Float!] schema in Float any-slice test
Align the FromAnySlice test schema with its name and its typed-slice
sibling ([Float!]) instead of duplicating the [Float64!] path.

+++ tests/integration/mutation/add/field_kinds/field_kind_float_test.go
@@ -458,7 +458,7 @@ func TestMutationAddFieldKinds_WithNonNillableFloatArray_FromAnySlice(t *testing
 			&action.AddCollection{
 				SDL: `
 					type User {
-						points: [Float64!]
+						points: [Float!]
 					}
 				`,
 			},
